refactor: 불필요한 오버라이딩 삭제
