+++ src/main/java/roomescape/domain/reservationdetail/ThemeRepository.java
@@ -2,15 +2,11 @@ package roomescape.domain.reservationdetail;
 
 import java.util.List;
 import java.util.Optional;
-import roomescape.exception.theme.NotFoundThemeException;
 
 public interface ThemeRepository {
     Theme save(Theme theme);
 
-    default Theme getById(Long id) {
-        return findById(id)
-                .orElseThrow(NotFoundThemeException::new);
-    }
+    Theme getById(Long id);
 
     Optional<Theme> findById(Long id);
