+++ src/main/java/roomescape/infra/repository/ReservationTimeJpaRepository.java
@@ -3,25 +3,22 @@ package roomescape.infra.repository;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
-import java.util.Optional;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.Repository;
 import org.springframework.data.repository.query.Param;
 import roomescape.domain.reservationdetail.ReservationTime;
 import roomescape.domain.reservationdetail.ReservationTimeRepository;
+import roomescape.exception.time.NotFoundReservationTimeException;
 
 public interface ReservationTimeJpaRepository extends
         ReservationTimeRepository,
         Repository<ReservationTime, LocalTime> {
 
     @Override
-    ReservationTime save(ReservationTime time);
-
-    @Override
-    Optional<ReservationTime> findById(Long id);
-
-    @Override
-    List<ReservationTime> findAll();
+    default ReservationTime getById(Long id) {
+        return findById(id)
+                .orElseThrow(NotFoundReservationTimeException::new);
+    }
 
     @Override
     @Query("""
@@ -38,10 +35,4 @@ public interface ReservationTimeJpaRepository extends
             @Param("date") LocalDate date,
             @Param("themeId") Long themeId
     );
-
-    @Override
-    boolean existsByStartAt(LocalTime startAt);
-
-    @Override
-    void delete(ReservationTime time);
 }
